refactor: change parseInt() to parseFloat() for float input acceptance

+++ script.js
@@ -9,7 +9,7 @@ document.addEventListener("DOMContentLoaded", () => {
 
     // math functions
     function add(a, b) {
-        return parseInt(a) + parseInt(b);
+        return parseFloat(a) + parseFloat(b);
     }
 
     function subtract(a, b) {
